Play and seek now available
I completed the ability to to play, pause and seek the video using the
player. I have also used a very basic style of the player.

The index.html is a demo file, copy everything inside of the body tag if
you want to embed it on your own website.